Output links test, db config yaml, local properties yaml

+++ models/init.rb
@@ -1,22 +1,32 @@
 # encoding: utf-8
 require 'sequel'
+require 'yaml'
 
-DB = Sequel.connect(:adapter => 'mysql',
-                    :host => 'localhost',
-                    :user => 'root',
-                    :password => '',
-                    :database => 'docula')
+$db_config = YAML.load_file('models/db.yml')
 
-DB.drop_table :docsets
-DB.create_table :docsets do
-  primary_key :id
-  String :name
-  String :branch
-  String :fs_path
+DB = Sequel.connect(:adapter => $db_config['adapter'],
+                    :host => $db_config['host'],
+                    :user => $db_config['user'],
+                    :password => $db_config['password'],
+                    :database => $db_config['database'])
+
+if ($db_config['recreate'])
+  if (DB.table_exists?(:docsets))
+    DB.drop_table :docsets
+  end
+
+  DB.create_table :docsets do
+    primary_key :id
+    String :name
+    String :branch
+    String :fs_path
+  end
+
+  docsets = DB[:docsets]
+
+  $db_config['data'].each do |record|
+    docsets.insert(:name => record['name'], :branch => record['branch'], :fs_path => record['fs_path'])
+  end
 end
-
-docsets = DB[:docsets]
-
-docsets.insert(:name => 'docula-sample', :branch => 'master', :fs_path => '/Users/Andre/GitHub/docula-sample')
 
 require_relative 'docset'
